+++ .gitignore
@@ -16,3 +16,5 @@ debug.test
 /rest-client.env.json
 .DS_Store
 .mongo
+remark42
+/bin/
