+++ Dockerfile
@@ -58,7 +58,7 @@ RUN \
     echo "runs outside of drone" && version="local"; \
     else version=${DRONE_TAG}${DRONE_BRANCH}${DRONE_PULL_REQUEST}-${DRONE_COMMIT:0:7}-$(date +%Y%m%d-%H:%M:%S); fi && \
     echo "version=$version" && \
-    go build -o remark -ldflags "-X main.revision=${version} -s -w" ./app
+    go build -o remark42 -ldflags "-X main.revision=${version} -s -w" ./app
 
 
 FROM node:10.6-alpine as build-frontend-deps
@@ -88,16 +88,19 @@ FROM umputun/baseimage:app-latest
 
 WORKDIR /srv
 
-ADD backend/scripts/*.sh /srv/
-ADD start.sh /srv/start.sh
-RUN chmod +x /srv/*.sh
+ADD entrypoint.sh /entrypoint.sh
+ADD backend/scripts/backup.sh /usr/local/bin/backup
+ADD backend/scripts/restore.sh /usr/local/bin/restore
+ADD backend/scripts/import.sh /usr/local/bin/import
+RUN chmod +x /entrypoint.sh /usr/local/bin/backup /usr/local/bin/restore /usr/local/bin/import
 
-COPY --from=build-backend /go/src/github.com/umputun/remark/backend/remark /srv/
+COPY --from=build-backend /go/src/github.com/umputun/remark/backend/remark42 /srv/remark42
 COPY --from=build-frontend /srv/web/public/ /srv/web
 RUN chown -R app:app /srv
+RUN ln -s /srv/remark42 /usr/bin/remark42
 
 EXPOSE 8080
 HEALTHCHECK --interval=30s --timeout=3s CMD curl --fail http://localhost:8080/ping || exit 1
 
-CMD ["/srv/start.sh"]
-ENTRYPOINT ["/init.sh"]
+CMD ["server"]
+ENTRYPOINT ["/entrypoint.sh"]
