+++ backend/app/cmd/testdata/import.txt
@@ -1,0 +1,3 @@
+blah
+blah2
+12345678
