+++ backend/scripts/backup.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -e
+/srv/remark42 backup $@
